Add event search in GIDD dashboard
